feat(connectCore): remove QR dependencies from connectCore
